+++ src/main/java/com/plivo/api/PlivoAPIService.java
@@ -36,6 +36,7 @@ import com.plivo.api.models.brand.Brand;
 import com.plivo.api.models.brand.BrandCreateResponse;
 import com.plivo.api.models.brand.BrandCreator;
 import com.plivo.api.models.brand.BrandUsecase;
+import com.plivo.api.models.brand.BrandDeleteResponse;
 import com.plivo.api.models.campaign.*;
 import com.plivo.api.models.profile.*;
 import com.plivo.api.models.token.TokenCreateResponse;
@@ -256,6 +257,9 @@ public interface PlivoAPIService {
   @GET("Account/{authId}/10dlc/Brand/{id}/")
   Call<Brand> brandGet(@Path("authId") String authId, @Path("id") String brandId);
 
+  @DELETE("Account/{authId}/10dlc/Brand/{id}/")
+  Call<BrandDeleteResponse> brandDelete(@Path("authId") String authId, @Path("id") String brandId);
+
   @GET("Account/{authId}/10dlc/Brand/{id}/usecases/")
   Call<BrandUsecase> brandUsecaseGet(@Path("authId") String authId, @Path("id") String brandId);
 
@@ -304,6 +308,9 @@ public interface PlivoAPIService {
 
   @GET("Account/{authId}/10dlc/Campaign/{campaignId}/")
   Call<Campaign> campaignGet(@Path("authId") String authId, @Path("campaignId") String campaignId);
+
+  @DELETE("Account/{authId}/10dlc/Campaign/{campaignId}/")
+  Call<CampaignDeleteResponse> campaignDelete(@Path("authId") String authId, @Path("campaignId") String campaignId);
 
   // Recording