+++ src/main/java/com/plivo/api/models/campaign/CampaignDeleter.java
@@ -1,0 +1,18 @@
+package com.plivo.api.models.campaign;
+
+import com.plivo.api.models.base.MessagingDeleter;
+import retrofit2.Call;
+
+public class CampaignDeleter extends MessagingDeleter<CampaignDeleteResponse> {
+
+  public CampaignDeleter(String id) {
+    super(id);
+    this.id = id;
+  }
+
+  @Override
+  protected Call<CampaignDeleteResponse> obtainCall() {
+    return client().getApiService().campaignDelete(client().getAuthId(), id);
+  }
+}
+